requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.6
django-cloudinary-storage==0.3.0
django-redis==6.0.0
django-unfold==0.67.0
djangorestframework==3.16.1
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
